# mcp-agents/contracts/mcp-agent-openapi.yaml
openapi: 3.1.0
info:
  title: NGKG MCP and Agent REST API
  version: 0.8.0
  description: Complete REST parity for NGKG semantic MCP tools, long-input compilation, managed agents, tenant tools, evidence-bound
    memory and distributed CPU qualification. Load this endpoint directly in Swagger UI.
servers:
- url: /
security:
- bearerAuth: []
...
paths:
...
  /v1/memories:
    post:
      tags:
      - Agent Memory
      operationId: proposeMemory
      x-mcp-tool: ngkg_memory_propose
      summary: Propose immutable working
      episodic: null
      semantic: null
      procedural or evidence memory: null
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryProposal'
      responses:
        '201':
          $ref: '#/components/responses/MemoryView'
        '422':
          description: Poisoning or evidence rejected
      description: This operation implements the documented "Propose immutable working" API capability. Use it when an API
        client needs this capability through the supported contract instead of coupling to an internal service. The request
        body uses application/json and the schema shown below; success returns the MemoryView response schema. Authentication,
        tenant authorization, resource ceilings and documented failure responses are enforced before a result is accepted.
  /v1/memories/search:
    post:
      tags:
      - Agent Memory
      operationId: searchMemories
      x-mcp-tool: ngkg_memory_search
      summary: Search only authorized
      current and unexpired memory: null
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemorySearch'
      responses:
        '200':
          description: Memory array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemoryView'
      description: This operation implements the documented "Search only authorized" API capability. Use it when an API client
        needs this capability through the supported contract instead of coupling to an internal service. The request body
        uses application/json and the schema shown below; success returns memory array. Authentication, tenant authorization,
        resource ceilings and documented failure responses are enforced before a result is accepted.
  /v1/memories/{memoryId}:
    get:
      tags:
      - Agent Memory
      operationId: getMemory
      summary: Read an exact authorized memory version
      parameters:
      - $ref: '#/components/parameters/memoryId'
      responses:
        '200':
          $ref: '#/components/responses/MemoryView'
      description: This operation implements the documented "Read an exact authorized memory version" API capability. Use
        it when an agent workflow needs durable, evidence-bound memory without allowing a model to write authoritative facts
        directly. Inputs use the documented path, query and header parameters; success returns the MemoryView response schema.
        Authentication, tenant authorization, resource ceilings and documented failure responses are enforced before a result
        is accepted.
  /v1/memories/{memoryId}/explain:
    get:
      tags:
      - Agent Memory
      operationId: explainMemory
      x-mcp-tool: ngkg_memory_explain
      summary: Explain provenance
      transitions: null
      edges and inclusion/exclusion policy: null
      parameters:
      - $ref: '#/components/parameters/memoryId'
      responses:
        '200':
          description: Memory explanation
      description: This operation implements the documented "Explain provenance" API capability. Use it when an agent workflow
        needs durable, evidence-bound memory without allowing a model to write authoritative facts directly. Inputs use the
        documented path, query and header parameters; success returns memory explanation. Authentication, tenant authorization,
        resource ceilings and documented failure responses are enforced before a result is accepted.
  /v1/memories/{memoryId}/validate:
    post:
      tags:
      - Agent Memory
      operationId: validateMemory
      x-mcp-tool: ngkg_memory_validate
      summary: Structurally validate memory and re-entail semantic statements through NGKG
      parameters:
      - $ref: '#/components/parameters/memoryId'
      responses:
        '200':
          description: Validation and proof evidence
        '422':
          description: Unknown or evidence mismatch
      description: This operation implements the documented "Structurally validate memory and re-entail semantic statements
        through NGKG" API capability. Use it when an agent workflow needs durable, evidence-bound memory without allowing
        a model to write authoritative facts directly. Inputs use the documented path, query and header parameters; success
        returns validation and proof evidence. Authentication, tenant authorization, resource ceilings and documented failure
        responses are enforced before a result is accepted.
  /v1/memories/{memoryId}/approve:
    post:
      tags:
      - Agent Memory
      operationId: approveMemory
      x-mcp-tool: ngkg_memory_approve
      summary: Approve an entailed semantic memory for publication activation
      parameters:
      - $ref: '#/components/parameters/memoryId'
      requestBody:
        $ref: '#/components/requestBodies/Reason'
      responses:
        '200':
          $ref: '#/components/responses/MemoryView'
      description: This operation implements the documented "Approve an entailed semantic memory for publication activation"
        API capability. Use it when an agent workflow needs durable, evidence-bound memory without allowing a model to write
        authoritative facts directly. The request body uses the documented media type and the schema shown below; success
        returns the MemoryView response schema. Authentication, tenant authorization, resource ceilings and documented failure
        responses are enforced before a result is accepted.
  /v1/memories/{memoryId}/publish:
    post:
      tags:
      - Agent Memory
      operationId: publishMemory
      x-mcp-tool: ngkg_memory_publish
      summary: Activate semantic memory only after re-entailment in an atomically published NGKG snapshot
      parameters:
      - $ref: '#/components/parameters/memoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryPublication'
      responses:
        '200':
          description: Publication receipt
        '422':
          description: Not entailed; no activation
      description: This operation implements the documented "Activate semantic memory only after re-entailment in an atomically
        published NGKG snapshot" API capability. Use it when an agent workflow needs durable, evidence-bound memory without
        allowing a model to write authoritative facts directly. The request body uses application/json and the schema shown
        below; success returns publication receipt. Authentication, tenant authorization, resource ceilings and documented
        failure responses are enforced before a result is accepted.
  /v1/memories/{memoryId}/supersede:
    post:
      tags:
      - Agent Memory
      operationId: supersedeMemory
      x-mcp-tool: ngkg_memory_supersede
      summary: Exclude an older memory using an immutable supersession edge
      parameters:
      - $ref: '#/components/parameters/memoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemorySupersede'
      responses:
        '200':
          $ref: '#/components/responses/MemoryView'
      description: This operation implements the documented "Exclude an older memory using an immutable supersession edge"
        API capability. Use it when an agent workflow needs durable, evidence-bound memory without allowing a model to write
        authoritative facts directly. The request body uses application/json and the schema shown below; success returns the
        MemoryView response schema. Authentication, tenant authorization, resource ceilings and documented failure responses
        are enforced before a result is accepted.
  /v1/memories/{memoryId}/revoke:
    post:
      tags:
      - Agent Memory
      operationId: revokeMemory
      x-mcp-tool: ngkg_memory_revoke
      summary: Revoke and exclude a memory without deleting history
      parameters:
      - $ref: '#/components/parameters/memoryId'
      requestBody:
        $ref: '#/components/requestBodies/Reason'
      responses:
        '200':
          $ref: '#/components/responses/MemoryView'
      description: This operation implements the documented "Revoke and exclude a memory without deleting history" API capability.
        Use it when an agent workflow needs durable, evidence-bound memory without allowing a model to write authoritative
        facts directly. The request body uses the documented media type and the schema shown below; success returns the MemoryView
        response schema. Authentication, tenant authorization, resource ceilings and documented failure responses are enforced
        before a result is accepted.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    memoryId:
      in: path
      name: memoryId
      required: true
      schema:
        type: string
        format: uuid
...
  requestBodies:
    Reason:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MemoryReason'
  responses:
...
    MemoryView:
      description: Exact authorized memory
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MemoryView'
  schemas:
...
    MemoryClass:
      enum:
      - WORKING
      - EPISODIC
      - SEMANTIC
      - PROCEDURAL
      - EVIDENCE
    MemoryState:
      enum:
      - PROPOSED
      - VALIDATING
      - VALIDATED
      - ENTAILED
      - CONTRADICTED
      - UNKNOWN
      - APPROVAL_REQUIRED
      - APPROVED
      - PUBLISHED
      - SUPERSEDED
      - REVOKED
      - REJECTED
      - EXPIRED
    Hash:
      type: string
      pattern: ^[0-9a-f]{64}$
    MemoryProposal:
      type: object
      additionalProperties: false
      required:
      - memoryClass
      - audience
      - contentType
      - content
      - provenance
      - retentionDays
      - legalHold
      - idempotencyKey
      properties:
        memoryClass:
          $ref: '#/components/schemas/MemoryClass'
        audience:
          enum:
          - OWNER
          - TENANT
        contentType:
          enum:
          - text/plain
          - application/json
          - application/n-triples
        content:
          type: string
          minLength: 1
          maxLength: 1048576
        sourceExecutionId:
          type:
          - string
          - 'null'
          format: uuid
        datasetId:
          type:
          - string
          - 'null'
          format: uuid
        snapshotId:
          type:
          - string
          - 'null'
          format: uuid
        authorizedGraphSetSha256:
          oneOf:
          - $ref: '#/components/schemas/Hash'
          - type: 'null'
        semanticResultSha256:
          oneOf:
          - $ref: '#/components/schemas/Hash'
          - type: 'null'
        answerCertificateSha256:
          oneOf:
          - $ref: '#/components/schemas/Hash'
          - type: 'null'
        provenance:
          type: object
        retentionDays:
          type: integer
          minimum: 1
          maximum: 3650
        legalHold:
          type: boolean
        expiresAtEpochMs:
          type:
          - integer
          - 'null'
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 256
    MemoryView:
      type: object
      additionalProperties: false
      required:
      - memoryId
      - memoryClass
      - audience
      - state
      - stateVersion
      - version
      - ownerSubject
      - contentType
      - content
      - contentSha256
      - provenance
      - retentionDays
      - legalHold
      - createdAtEpochMs
      - updatedAtEpochMs
      properties:
        memoryId:
          type: string
          format: uuid
        memoryClass:
          $ref: '#/components/schemas/MemoryClass'
        audience:
          enum:
          - OWNER
          - TENANT
        state:
          $ref: '#/components/schemas/MemoryState'
        stateVersion:
          type: integer
        version:
          type: integer
        ownerSubject:
          type: string
        contentType:
          enum:
          - text/plain
          - application/json
          - application/n-triples
        content:
          type: string
        contentSha256:
          $ref: '#/components/schemas/Hash'
        sourceExecutionId:
          type:
          - string
          - 'null'
          format: uuid
        datasetId:
          type:
          - string
          - 'null'
          format: uuid
        snapshotId:
          type:
          - string
          - 'null'
          format: uuid
        authorizedGraphSetSha256:
          oneOf:
          - $ref: '#/components/schemas/Hash'
          - type: 'null'
        semanticResultSha256:
          oneOf:
          - $ref: '#/components/schemas/Hash'
          - type: 'null'
        answerCertificateSha256:
          oneOf:
          - $ref: '#/components/schemas/Hash'
          - type: 'null'
        provenance:
          type: object
        retentionDays:
          type: integer
        legalHold:
          type: boolean
        expiresAtEpochMs:
          type:
          - integer
          - 'null'
        createdAtEpochMs:
          type: integer
        updatedAtEpochMs:
          type: integer
    MemorySearch:
      type: object
      additionalProperties: false
      properties:
        query:
          type: string
          maxLength: 4096
        memoryClass:
          $ref: '#/components/schemas/MemoryClass'
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
    MemoryReason:
      type: object
      additionalProperties: false
      required:
      - reasonCode
      properties:
        reasonCode:
          type: string
          pattern: ^[A-Z0-9_]{1,128}$
    MemoryPublication:
      type: object
      additionalProperties: false
      required:
      - ngkgOperationId
      - publishedSnapshotId
      properties:
        ngkgOperationId:
          type: string
          format: uuid
        publishedSnapshotId:
          type: string
          format: uuid
    MemorySupersede:
      type: object
      additionalProperties: false
      required:
      - supersedingMemoryId
      - reasonCode
      properties:
        supersedingMemoryId:
          type: string
          format: uuid
        reasonCode:
          type: string
          pattern: ^[A-Z0-9_]{1,128}$
...